src/requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
astroid==2.2.2
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-oauth-toolkit==1.2.0
django-sslserver==0.22
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.10
jedi==0.15.2
Jinja2==2.10.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.2
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
pip-review==1.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pylint==2.3.1
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.6.0
readme-renderer==24.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.13.0
sqlparse==0.3.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-object-detection-api==0.1.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
twine==3.1.1
typed-ast==1.3.1
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.1
zipp==0.6.0
